# Add a fast path for `optimize_projection`