@alamb 2025-04-17
🤖 `./gh_compare_branch.sh` [Benchmark Script](https://github.com/alamb/datafusion-benchmarking/blob/main/gh_compare_branch.sh) Running
Linux aal-dev 6.8.0-1016-gcp #18-Ubuntu SMP Fri Oct  4 22:16:29 UTC 2024 x86_64 x86_64 x86_64 GNU/Linux
Comparing speed_up_optimize_projection (f62173e76fb4dc5e6973a88528fee0924d9ec41b) to 2cba3ad17874b742868123600717a10fdb55ec36 [diff](https://github.com/apache/datafusion/compare/2cba3ad17874b742868123600717a10fdb55ec36..f62173e76fb4dc5e6973a88528fee0924d9ec41b)
Benchmarks: tpch_mem clickbench_partitioned clickbench_extended
Results will be posted here when complete